+++ src/e3j/pallas_ops/tensor_product/__init__.py
@@ -1,0 +1,29 @@
+# Copyright (c) 2026 InstaDeep Ltd
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from e3j.pallas_ops.tensor_product.common.params import PallasTensorProductParams
+from e3j.pallas_ops.tensor_product.mosaic_tpu import tensor_product_pallas_mosaic_tpu
+from e3j.pallas_ops.tensor_product.mosaic_tpu.params import (
+    PallasMosaicTPUTensorProductParams,
+    PallasMosaicTPUTensorProductParamsBwdConfig,
+    PallasMosaicTPUTensorProductParamsFwdConfig,
+)
+
+__all__ = [
+    "PallasTensorProductParams",
+    "PallasMosaicTPUTensorProductParams",
+    "PallasMosaicTPUTensorProductParamsFwdConfig",
+    "PallasMosaicTPUTensorProductParamsBwdConfig",
+    "tensor_product_pallas_mosaic_tpu",
+]